# frontend/generated/backend-openapi.json
{
  "components": {
    "schemas": {
      "ApiKeyTestResult": {
        "description": "Result of testing an API key for a remote backend.",
        "properties": {
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid"
        ],
        "title": "ApiKeyTestResult",
        "type": "object"
      },
...
      "BackendDescriptor": {
        "description": "Describes a captioning backend and whether it is ready to use.",
        "properties": {
          "backend_id": {
            "enum": [
              "local",
              "gemini",
              "openai",
              "anthropic",
              "openai_compatible"
            ],
            "title": "Backend Id",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "is_configured": {
            "title": "Is Configured",
            "type": "boolean"
          },
          "is_local": {
            "title": "Is Local",
            "type": "boolean"
          }
        },
        "required": [
          "backend_id",
          "display_name",
          "is_configured",
          "is_local"
        ],
        "title": "BackendDescriptor",
        "type": "object"
      },
      "CaptionBatchRequest": {
        "properties": {
          "backend_id": {
            "default": "local",
            "enum": [
              "local",
              "gemini",
              "openai",
              "anthropic",
              "openai_compatible"
            ],
            "title": "Backend Id",
            "type": "string"
          },
          "clip_ids": {
            "items": {
              "type": "string"
            },
            "title": "Clip Ids",
            "type": "array"
          },
          "dataset_dir": {
            "title": "Dataset Dir",
            "type": "string"
          },
          "prompt_template": {
            "$ref": "#/components/schemas/PromptTemplate",
            "default": {
              "frame_count": 8,
              "system_prompt": "You are a video annotation assistant. Describe the subject, action, framing, expression, and setting in one to three sentences. Use video terminology (e.g., \"shot\", \"footage\"). Do not use words like \"photograph\" or \"still\". Do not invent details not visible in the frames. Begin directly with the subject. Do not open the caption with a preamble such as \"The video clip features\", \"This video shows\", \"The image depicts\", or any similar framing phrase.",
              "user_prompt": "Describe this short video clip."
            }
          }
        },
        "required": [
          "dataset_dir",
          "clip_ids"
        ],
        "title": "CaptionBatchRequest",
        "type": "object"
      },
      "CaptionBatchStatus": {
        "description": "Status of a batch captioning job.",
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "integer"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CaptionResult"
            },
            "title": "Results",
            "type": "array"
          },
          "state": {
            "enum": [
              "running",
              "complete",
              "cancelled",
              "error"
            ],
            "title": "State",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "state",
          "total",
          "completed",
          "failed",
          "results"
        ],
        "title": "CaptionBatchStatus",
        "type": "object"
      },
      "CaptionClipRequest": {
        "properties": {
          "backend_id": {
            "default": "local",
            "enum": [
              "local",
              "gemini",
              "openai",
              "anthropic",
              "openai_compatible"
            ],
            "title": "Backend Id",
            "type": "string"
          },
          "clip_id": {
            "title": "Clip Id",
            "type": "string"
          },
          "dataset_dir": {
            "title": "Dataset Dir",
            "type": "string"
          },
          "prompt_template": {
            "$ref": "#/components/schemas/PromptTemplate",
            "default": {
              "frame_count": 8,
              "system_prompt": "You are a video annotation assistant. Describe the subject, action, framing, expression, and setting in one to three sentences. Use video terminology (e.g., \"shot\", \"footage\"). Do not use words like \"photograph\" or \"still\". Do not invent details not visible in the frames. Begin directly with the subject. Do not open the caption with a preamble such as \"The video clip features\", \"This video shows\", \"The image depicts\", or any similar framing phrase.",
              "user_prompt": "Describe this short video clip."
            }
          }
        },
        "required": [
          "dataset_dir",
          "clip_id"
        ],
        "title": "CaptionClipRequest",
        "type": "object"
      },
      "CaptionResult": {
        "description": "Result of captioning a single clip.",
        "properties": {
          "backend_used": {
            "enum": [
              "local",
              "gemini",
              "openai",
              "anthropic",
              "openai_compatible"
            ],
            "title": "Backend Used",
            "type": "string"
          },
          "caption": {
            "title": "Caption",
            "type": "string"
          },
          "clip_id": {
            "title": "Clip Id",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "clip_id",
          "caption",
          "backend_used",
          "success"
        ],
        "title": "CaptionResult",
        "type": "object"
      },
...
      "HTTPErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "HTTPErrorResponse",
        "type": "object"
      },
...
      "LocalModelChoice": {
        "description": "User's choice of local VLM model.",
        "properties": {
          "abliterated": {
            "default": false,
            "title": "Abliterated",
            "type": "boolean"
          },
          "family": {
            "const": "qwen3-vl",
            "default": "qwen3-vl",
            "title": "Family",
            "type": "string"
          },
          "quantization": {
            "default": "fp16",
            "enum": [
              "fp16",
              "8bit",
              "4bit"
            ],
            "title": "Quantization",
            "type": "string"
          },
          "size": {
            "default": "4B",
            "enum": [
              "2B",
              "4B",
              "8B",
              "32B"
            ],
            "title": "Size",
            "type": "string"
          }
        },
        "title": "LocalModelChoice",
        "type": "object"
      },
...
      "ModelSetupStatus": {
        "description": "Status of local model setup (download / loading / ready / error).",
        "properties": {
          "current_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current File"
          },
          "downloaded_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Downloaded Bytes"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "model_choice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalModelChoice"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "default": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "state": {
            "enum": [
              "not_started",
              "downloading",
              "loading",
              "ready",
              "error"
            ],
            "title": "State",
            "type": "string"
          },
          "total_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Bytes"
          }
        },
        "required": [
          "state"
        ],
        "title": "ModelSetupStatus",
        "type": "object"
      },
...
      "PromptTemplate": {
        "description": "System + user prompt for captioning.",
        "properties": {
          "frame_count": {
            "default": 8,
            "title": "Frame Count",
            "type": "integer"
          },
          "system_prompt": {
            "default": "You are a video annotation assistant. Describe the subject, action, framing, expression, and setting in one to three sentences. Use video terminology (e.g., \"shot\", \"footage\"). Do not use words like \"photograph\" or \"still\". Do not invent details not visible in the frames. Begin directly with the subject. Do not open the caption with a preamble such as \"The video clip features\", \"This video shows\", \"The image depicts\", or any similar framing phrase.",
            "title": "System Prompt",
            "type": "string"
          },
          "user_prompt": {
            "default": "Describe this short video clip.",
            "title": "User Prompt",
            "type": "string"
          }
        },
        "title": "PromptTemplate",
        "type": "object"
      },
...
      "SaveApiKeyRequest": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          }
        },
        "required": [
          "key"
        ],
        "title": "SaveApiKeyRequest",
        "type": "object"
      },
...
      "SelectLocalModelRequest": {
        "properties": {
          "choice": {
            "$ref": "#/components/schemas/LocalModelChoice"
          },
          "gpu_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Index"
          }
        },
        "required": [
          "choice"
        ],
        "title": "SelectLocalModelRequest",
        "type": "object"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenLTX Trainer Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/caption/api-keys/{provider}": {
      "delete": {
        "operationId": "route_delete_api_key_api_caption_api_keys__provider__delete",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "enum": [
                "local",
                "gemini",
                "openai",
                "anthropic",
                "openai_compatible"
              ],
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Delete Api Key",
        "tags": [
          "caption"
        ]
      },
      "post": {
        "operationId": "route_save_api_key_api_caption_api_keys__provider__post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "enum": [
                "local",
                "gemini",
                "openai",
                "anthropic",
                "openai_compatible"
              ],
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Save Api Key",
        "tags": [
          "caption"
        ]
      }
    },
    "/api/caption/api-keys/{provider}/test": {
      "post": {
        "operationId": "route_test_api_key_api_caption_api_keys__provider__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "enum": [
                "local",
                "gemini",
                "openai",
                "anthropic",
                "openai_compatible"
              ],
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyTestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Test Api Key",
        "tags": [
          "caption"
        ]
      }
    },
    "/api/caption/backends": {
      "get": {
        "operationId": "route_list_backends_api_caption_backends_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BackendDescriptor"
                  },
                  "title": "Response Route List Backends Api Caption Backends Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route List Backends",
        "tags": [
          "caption"
        ]
      }
    },
    "/api/caption/batch": {
      "post": {
        "operationId": "route_caption_batch_api_caption_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptionBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptionBatchStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Caption Batch",
        "tags": [
          "caption"
        ]
      }
    },
    "/api/caption/clip": {
      "post": {
        "operationId": "route_caption_clip_api_caption_clip_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptionClipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Caption Clip",
        "tags": [
          "caption"
        ]
      }
    },
    "/api/caption/jobs/{job_id}": {
      "get": {
        "operationId": "route_get_batch_status_api_caption_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CaptionBatchStatus"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Route Get Batch Status Api Caption Jobs  Job Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Get Batch Status",
        "tags": [
          "caption"
        ]
      }
    },
    "/api/caption/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "route_cancel_batch_api_caption_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Cancel Batch",
        "tags": [
          "caption"
        ]
      }
    },
    "/api/caption/local-model/choices": {
      "get": {
        "operationId": "route_list_local_model_choices_api_caption_local_model_choices_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LocalModelChoice"
                  },
                  "title": "Response Route List Local Model Choices Api Caption Local Model Choices Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route List Local Model Choices",
        "tags": [
          "caption"
        ]
      }
    },
    "/api/caption/local-model/select": {
      "post": {
        "operationId": "route_select_local_model_api_caption_local_model_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectLocalModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSetupStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Select Local Model",
        "tags": [
          "caption"
        ]
      }
    },
    "/api/caption/local-model/setup-status": {
      "get": {
        "operationId": "route_get_local_model_status_api_caption_local_model_setup_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSetupStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Get Local Model Status",
        "tags": [
          "caption"
        ]
      }
    },
    "/api/caption/local-model/unload": {
      "post": {
        "operationId": "route_unload_local_model_api_caption_local_model_unload_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSetupStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Unload Local Model",
        "tags": [
          "caption"
        ]
      }
    },
...
  }
}